Update Play.js
Added some spiked obstacles and some that are rotating. Also added some "formations" or something like that so the obstacles wont just be a mess. I also tried adding a power up bar but thats not working yet. Basically the only things that works is the spiked obstacles and the rotating ones (including the "normal" obstacles). Thats all that works ;-;